test: replace 51 inline browser assert blocks with one shared module (#777)
Every browser test hand-rolled its own const assert = {...} (node:assert is
absent in the browser), duplicated across ~51 files and drifted (some had a
Map-aware deepEqual, some a plain JSON one). Extract one shared
test/browser-assert.js, a superset of every method the inline blocks defined
(ok/equal/deepEqual/strictEqual/notStrictEqual/notEqual/match/isTrue/isFalse/
isOk/notOk/isNaN/isUndefined/isArray/doesNotThrow/throws), matching their
semantics exactly (strict !== equal, JSON-stringify deepEqual with Map + NaN
normalization + sorted keys, sync doesNotThrow, async throws). Each browser
test imports it via a computed relative path. Green on Chromium, Firefox, and
WebKit.
