+++ packages/cli/templates/test/hello/browser/hello.test.js
@@ -17,10 +17,7 @@
 import { html } from '@webjsdev/core';
 import { ssrFixture, assertNoA11yViolations } from '@webjsdev/core/testing';
 
-const assert = {
-  ok: (v, msg) => { if (!v) throw new Error(msg || `Expected truthy`); },
-  equal: (a, b, msg) => { if (a !== b) throw new Error(msg || `Expected ${b}, got ${a}`); },
-};
+import { assert } from '../../../../../../test/browser-assert.js';
 
 suite('Example browser tests', () => {
   test('DOM is real (not jsdom/linkedom)', () => {
