+++ packages/core/test/directives/browser/directives-async-stream.test.js
@@ -27,10 +27,7 @@ import { html } from '../../../src/html.js';
 import { render } from '../../../src/render-client.js';
 import { asyncAppend, asyncReplace } from '../../../src/directives.js';
 
-const assert = {
-  ok: (v, msg) => { if (!v) throw new Error(msg || `Expected truthy, got ${v}`); },
-  equal: (a, b, msg) => { if (a !== b) throw new Error(msg || `Expected ${JSON.stringify(b)}, got ${JSON.stringify(a)}`); },
-};
+import { assert } from '../../../../../test/browser-assert.js';
 
 const nextFrame = () => new Promise((r) => requestAnimationFrame(() => r()));
