+++ packages/core/test/directives/browser/directives.test.js
@@ -15,12 +15,7 @@ import {
   cache, until, asyncAppend, asyncReplace,
 } from '../../../src/directives.js';
 
-const assert = {
-  ok: (v, msg) => { if (!v) throw new Error(msg || `Expected truthy, got ${v}`); },
-  equal: (a, b, msg) => { if (a !== b) throw new Error(msg || `Expected ${JSON.stringify(b)}, got ${JSON.stringify(a)}`); },
-  notStrictEqual: (a, b, msg) => { if (a === b) throw new Error(msg || 'Expected different references'); },
-  strictEqual: (a, b, msg) => { if (a !== b) throw new Error(msg || 'Expected strict equal'); },
-};
+import { assert } from '../../../../../test/browser-assert.js';
 
 suite('Directives in a real browser', () => {
