+++ packages/core/test/lifecycle/browser/component-lifecycle.test.js
@@ -12,15 +12,7 @@
 import { html } from '../../../src/html.js';
 import { WebComponent, prop } from '../../../src/component.js';
 
-const assert = {
-  ok: (v, msg) => { if (!v) throw new Error(msg || `Expected truthy, got ${v}`); },
-  equal: (a, b, msg) => { if (a !== b) throw new Error(msg || `Expected ${JSON.stringify(b)}, got ${JSON.stringify(a)}`); },
-  deepEqual: (a, b, msg) => {
-    if (JSON.stringify(a) !== JSON.stringify(b)) {
-      throw new Error(msg || `deepEqual failed: ${JSON.stringify(a)} !== ${JSON.stringify(b)}`);
-    }
-  },
-};
+import { assert } from '../../../../../test/browser-assert.js';
 
 suite('Lifecycle hooks in a real browser', () => {
