+++ packages/core/test/lifecycle/browser/controllers-port.test.js
@@ -11,15 +11,7 @@
 import { html } from '../../../src/html.js';
 import { WebComponent } from '../../../src/component.js';
 
-const assert = {
-  ok: (v, msg) => { if (!v) throw new Error(msg || `Expected truthy, got ${v}`); },
-  equal: (a, b, msg) => { if (a !== b) throw new Error(msg || `Expected ${JSON.stringify(b)}, got ${JSON.stringify(a)}`); },
-  deepEqual: (a, b, msg) => {
-    if (JSON.stringify(a) !== JSON.stringify(b)) {
-      throw new Error(msg || `deepEqual failed: ${JSON.stringify(a)} !== ${JSON.stringify(b)}`);
-    }
-  },
-};
+import { assert } from '../../../../../test/browser-assert.js';
 
 let _uid = 0;
 function uniqTag(prefix) {
