+++ packages/core/test/routing/browser/data-navigating-optin.test.js
@@ -12,10 +12,7 @@
  */
 import { enableClientRouter, navigate } from '../../../src/router-client.js';
 
-const assert = {
-  ok: (v, msg) => { if (!v) throw new Error(msg || `Expected truthy, got ${v}`); },
-  equal: (a, b, msg) => { if (a !== b) throw new Error(msg || `Expected ${JSON.stringify(b)}, got ${JSON.stringify(a)}`); },
-};
+import { assert } from '../../../../../test/browser-assert.js';
 
 const sleep = (ms) => new Promise((r) => setTimeout(r, ms));
