+++ packages/core/test/routing/browser/frame-src-self-load.test.js
@@ -20,10 +20,7 @@
 import { enableClientRouter } from '../../../src/router-client.js';
 import '../../../src/webjs-frame.js';
 
-const assert = {
-  ok: (v, msg) => { if (!v) throw new Error(msg || `Expected truthy, got ${v}`); },
-  equal: (a, b, msg) => { if (a !== b) throw new Error(msg || `Expected ${JSON.stringify(b)}, got ${JSON.stringify(a)}`); },
-};
+import { assert } from '../../../../../test/browser-assert.js';
 const tick = () => new Promise((r) => setTimeout(r, 0));
 async function settle() { for (let i = 0; i < 6; i++) await tick(); }
 /** Poll until `cond()` is truthy (or a generous timeout), for a step gated on
