+++ packages/core/test/routing/browser/nav-scroll-instant.test.js
@@ -13,10 +13,7 @@
  */
 import { enableClientRouter, navigate } from '../../../src/router-client.js';
 
-const assert = {
-  ok: (v, msg) => { if (!v) throw new Error(msg || `Expected truthy, got ${v}`); },
-  equal: (a, b, msg) => { if (a !== b) throw new Error(msg || `Expected ${JSON.stringify(b)}, got ${JSON.stringify(a)}`); },
-};
+import { assert } from '../../../../../test/browser-assert.js';
 
 suite('Client router: nav scroll is instant under scroll-behavior:smooth (#601)', () => {
   let origFetch, origScrollTo, calls;
