+++ packages/core/test/routing/browser/progressive-stream.test.js
@@ -13,10 +13,7 @@ import { WebComponent } from '../../../src/component.js';
 import { html } from '../../../src/html.js';
 
 const { suite, test } = window.Mocha ? Mocha : { suite, test };
-const assert = {
-  ok: (v, msg) => { if (!v) throw new Error(msg || `Expected truthy, got ${v}`); },
-  equal: (a, b, msg) => { if (a !== b) throw new Error(msg || `Expected ${b}, got ${a}`); },
-};
+import { assert } from '../../../../../test/browser-assert.js';
 const tick = (ms = 0) => new Promise((r) => setTimeout(r, ms));
 
 let host;
