+++ packages/core/test/routing/browser/stream-action.test.js
@@ -18,10 +18,7 @@ import '../../../src/webjs-stream.js';
 import { enableClientRouter } from '../../../src/router-client.js';
 import { connectWS } from '../../../src/websocket-client.js';
 
-const assert = {
-  ok: (v, msg) => { if (!v) throw new Error(msg || `Expected truthy, got ${v}`); },
-  equal: (a, b, msg) => { if (a !== b) throw new Error(msg || `Expected ${JSON.stringify(b)}, got ${JSON.stringify(a)}`); },
-};
+import { assert } from '../../../../../test/browser-assert.js';
 const tick = () => new Promise((r) => setTimeout(r, 0));
 async function settle() { for (let i = 0; i < 4; i++) await tick(); }
