+++ packages/core/test/signals/browser/signal-hydration.test.js
@@ -12,10 +12,7 @@ import { html } from '../../../src/html.js';
 import { WebComponent, prop } from '../../../src/component.js';
 import { signal, computed } from '../../../src/signal.js';
 
-const assert = {
-  ok: (v, msg) => { if (!v) throw new Error(msg || `Expected truthy, got ${v}`); },
-  equal: (a, b, msg) => { if (a !== b) throw new Error(msg || `Expected ${JSON.stringify(b)}, got ${JSON.stringify(a)}`); },
-};
+import { assert } from '../../../../../test/browser-assert.js';
 
 suite('Signal + SSR hydration roundtrip', () => {
   let next = 0;
