+++ packages/core/test/signals/browser/signal-slot-integration.test.js
@@ -18,10 +18,7 @@ import { WebComponent } from '../../../src/component.js';
 import { signal, computed } from '../../../src/signal.js';
 import { watch } from '../../../src/directives.js';
 
-const assert = {
-  ok: (v, msg) => { if (!v) throw new Error(msg || `Expected truthy, got ${v}`); },
-  equal: (a, b, msg) => { if (a !== b) throw new Error(msg || `Expected ${JSON.stringify(b)}, got ${JSON.stringify(a)}`); },
-};
+import { assert } from '../../../../../test/browser-assert.js';
 
 suite('signal + light-DOM slot integration', () => {
   let nextTag = 0;
