+++ packages/core/test/signals/browser/watch-directive.test.js
@@ -8,10 +8,7 @@ import { WebComponent } from '../../../src/component.js';
 import { watch } from '../../../src/directives.js';
 import { signal } from '../../../src/signal.js';
 
-const assert = {
-  ok: (v, msg) => { if (!v) throw new Error(msg || `Expected truthy, got ${v}`); },
-  equal: (a, b, msg) => { if (a !== b) throw new Error(msg || `Expected ${JSON.stringify(b)}, got ${JSON.stringify(a)}`); },
-};
+import { assert } from '../../../../../test/browser-assert.js';
 
 suite('watch() directive', () => {
   let nextTag = 0;
