+++ packages/core/test/suspense/browser/webjs-suspense-client.test.js
@@ -11,10 +11,7 @@ import { WebComponent } from '../../../src/component.js';
 import '../../../src/webjs-suspense.js';
 
 const { suite, test } = window.Mocha ? Mocha : { suite, test };
-const assert = {
-  ok: (v, msg) => { if (!v) throw new Error(msg || `Expected truthy, got ${v}`); },
-  equal: (a, b, msg) => { if (a !== b) throw new Error(msg || `Expected ${b}, got ${a}`); },
-};
+import { assert } from '../../../../../test/browser-assert.js';
 const tick = (ms = 0) => new Promise((r) => setTimeout(r, ms));
 
 let host;
