+++ packages/server/test/dev/browser/reload-worker.test.js
@@ -12,11 +12,7 @@
  */
 import { startReloadWorker } from '../../../src/dev-reload-worker.js';
 
-const assert = {
-  ok: (v, msg) => { if (!v) throw new Error(msg || `Expected truthy, got ${v}`); },
-  equal: (a, b, msg) => { if (a !== b) throw new Error(msg || `Expected ${JSON.stringify(b)}, got ${JSON.stringify(a)}`); },
-  deepEqual: (a, b, msg) => { if (JSON.stringify(a) !== JSON.stringify(b)) throw new Error(msg || `Expected ${JSON.stringify(b)}, got ${JSON.stringify(a)}`); },
-};
+import { assert } from '../../../../../test/browser-assert.js';
 
 class FakeEventSource {
   constructor(url) { this.url = url; this._l = {}; FakeEventSource.last = this; }
