+++ packages/ui/test/components/browser/ui-overlay.test.js
@@ -19,11 +19,7 @@
 import { html } from '../../../../core/src/html.js';
 import { render } from '../../../../core/src/render-client.js';
 
-const assert = {
-  ok: (v, msg) => { if (!v) throw new Error(msg || `Expected truthy, got ${v}`); },
-  equal: (a, b, msg) => { if (a !== b) throw new Error(msg || `Expected ${b}, got ${a}`); },
-  match: (s, re, msg) => { if (!re.test(s)) throw new Error(msg || `Expected ${s} to match ${re}`); },
-};
+import { assert } from '../../../../../test/browser-assert.js';
 
 const COMPONENTS_DIR = '/packages/ui/packages/registry/components';
