+++ website/test/components/browser/copy-cmd.test.js
@@ -13,10 +13,7 @@
  */
 import '#components/copy-cmd.ts';
 
-const assert = {
-  ok: (v, msg) => { if (!v) throw new Error(msg || `Expected truthy, got ${v}`); },
-  equal: (a, b, msg) => { if (a !== b) throw new Error(msg || `Expected ${JSON.stringify(b)}, got ${JSON.stringify(a)}`); },
-};
+import { assert } from '../../../../test/browser-assert.js';
 
 const tick = (ms = 0) => new Promise((r) => setTimeout(r, ms));
